requirements.txt
Flask~=1.1.2
PyJWT~=1.7.0
cryptography~=2.6.1
pip~=18.1
wheel~=0.32.3
keyring~=18.0.1
Mako~=1.0.7
MarkupSafe~=1.1.0
pytz~=2019.2
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
blinker~=1.4
python-dateutil~=2.7.3
six~=1.12.0
psycopg2~=2.8.5
simplejson~=3.16.0
setuptools~=41.1.0
future~=0.16.0
argcomplete~=1.8.1